requirements.txt
tensorflow==2.17.1
keras==3.5.0
numpy==1.26.2
mysql-connector-python==9.1.0
flask==3.1.0
werkzeug==3.1.3
